move log.py back
