+++ bentoml/configuration/__init__.py
@@ -135,7 +135,7 @@ def _reset_bentoml_home(new_bentoml_home_directory):
     _config = load_config()
 
     # re-config logging
-    from bentoml.configuration.logging import configure_logging
+    from bentoml import configure_logging
 
     root = logging.getLogger()
     map(root.removeHandler, root.handlers[:])
